refactor: unify per-feature widget dirs under views/
The codebase split feature widgets between views/ and components/ with no
consistent rule (channels/messaging used components/ even for full screens;
everyone else used views/). Standardize on views/ (the documented majority).

- channels/components/ -> channels/views/
- messaging/components/ -> messaging/views/ (incl. box/ subtree)
- user/components/self_status_button.dart -> user/views/
- remove empty notifications/components/ and voice/components/
- update all import paths in lib/ and test/ accordingly

Pure move + import-path rewrite; no code logic changed.
