+++ lib/features/messaging/views/box/accord_message_content.dart
@@ -6,8 +6,8 @@ import 'package:bonfire/features/channels/controllers/read_state.dart';
 import 'package:bonfire/features/channels/models/open_tab.dart';
 import 'package:bonfire/features/member/controllers/accord_members.dart';
 import 'package:bonfire/features/member/views/accord_member_popout.dart';
-import 'package:bonfire/features/messaging/components/box/accord_markdown_box.dart';
-import 'package:bonfire/features/messaging/components/box/accord_message_markup.dart';
+import 'package:bonfire/features/messaging/views/box/accord_markdown_box.dart';
+import 'package:bonfire/features/messaging/views/box/accord_message_markup.dart';
 import 'package:bonfire/features/messaging/controllers/accord_emojis.dart';
 import 'package:bonfire/features/server/controllers/connections.dart';
 import 'package:bonfire/features/settings/controllers/settings.dart';
