+++ lib/features/messaging/views/image_lightbox.dart
