+++ lib/features/spaces/views/accord_home.dart
@@ -6,8 +6,8 @@ import 'package:collection/collection.dart';
 import 'package:pasteboard/pasteboard.dart';
 import 'package:bonfire/features/authentication/models/accord_auth.dart';
 import 'package:bonfire/features/authentication/repositories/accord_auth.dart';
-import 'package:bonfire/features/channels/components/channel_context_menu.dart';
-import 'package:bonfire/features/channels/components/channel_management.dart';
+import 'package:bonfire/features/channels/views/channel_context_menu.dart';
+import 'package:bonfire/features/channels/views/channel_management.dart';
 import 'package:bonfire/features/channels/controllers/accord_channels.dart';
 import 'package:bonfire/features/channels/controllers/open_tabs.dart';
 import 'package:bonfire/features/channels/controllers/read_state.dart';
@@ -19,15 +19,15 @@ import 'package:bonfire/features/member/utils/member_display.dart';
 import 'package:bonfire/features/member/views/accord_member_avatar.dart';
 import 'package:bonfire/features/member/views/accord_member_list.dart';
 import 'package:bonfire/features/member/views/accord_member_popout.dart';
-import 'package:bonfire/features/messaging/components/box/accord_embed_box.dart';
-import 'package:bonfire/features/messaging/components/emoji_picker.dart';
-import 'package:bonfire/features/messaging/components/box/accord_message_content.dart';
-import 'package:bonfire/features/messaging/components/forum_view.dart';
-import 'package:bonfire/features/messaging/components/thread_view.dart';
-import 'package:bonfire/features/messaging/components/image_lightbox.dart';
-import 'package:bonfire/features/messaging/components/inline_audio_player.dart';
-import 'package:bonfire/features/messaging/components/inline_video_player.dart';
-import 'package:bonfire/features/messaging/components/pinned_messages.dart';
+import 'package:bonfire/features/messaging/views/box/accord_embed_box.dart';
+import 'package:bonfire/features/messaging/views/emoji_picker.dart';
+import 'package:bonfire/features/messaging/views/box/accord_message_content.dart';
+import 'package:bonfire/features/messaging/views/forum_view.dart';
+import 'package:bonfire/features/messaging/views/thread_view.dart';
+import 'package:bonfire/features/messaging/views/image_lightbox.dart';
+import 'package:bonfire/features/messaging/views/inline_audio_player.dart';
+import 'package:bonfire/features/messaging/views/inline_video_player.dart';
+import 'package:bonfire/features/messaging/views/pinned_messages.dart';
 import 'package:bonfire/features/messaging/controllers/accord_emojis.dart';
 import 'package:bonfire/features/messaging/controllers/accord_messages.dart';
 import 'package:bonfire/features/messaging/utils/emoji_catalog.dart';
@@ -48,7 +48,7 @@ import 'package:bonfire/features/spaces/views/accord_invites.dart';
 import 'package:bonfire/features/spaces/views/accord_reports.dart';
 import 'package:bonfire/features/spaces/utils/message_time.dart';
 import 'package:bonfire/features/spaces/views/accord_search.dart';
-import 'package:bonfire/features/user/components/self_status_button.dart';
+import 'package:bonfire/features/user/views/self_status_button.dart';
 import 'package:bonfire/features/user/views/accord_direct_messages.dart';
 import 'package:bonfire/features/settings/controllers/settings.dart';
 import 'package:bonfire/features/settings/models/accord_settings.dart';
